apps/upload-bridge/requirements.txt
# Upload Bridge Requirements
# Core dependencies
PySide6>=6.5.0
pathlib2>=2.3.7

# ESP chip support
esptool>=4.6.0

# Arduino CLI support
# Note: Arduino CLI needs to be installed separately
# Download from: https://arduino.github.io/arduino-cli/

# AVR chip support
# Note: avrdude and avr-gcc need to be installed separately
# Download from: https://www.microchip.com/en-us/tools-resources/develop/microchip-studio

# STM32 chip support
# Note: stm32flash and arm-none-eabi-gcc need to be installed separately
# Download from: https://developer.arm.com/downloads/-/gnu-rm

# PIC chip support
# Note: MPLAB X IDE and XC compilers need to be installed separately
# Download from: https://www.microchip.com/en-us/tools-resources/develop/mplab-x-ide

# Nuvoton chip support
# Note: Nu-Link tools need to be installed separately
# Download from: https://www.nuvoton.com/tool-and-software/development-tool-hardware/

# Additional utilities
pyserial>=3.5
pyyaml>=6.0
colorama>=0.4.4
tqdm>=4.64.0
requests>=2.25.0

# Media conversion support
Pillow>=10.0.0
opencv-python>=4.8.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.8
numpy>=1.24.0

# License system security
cryptography>=41.0.0
certifi>=2023.0.0

# Authentication system (standard library - no pip install needed)
# - webbrowser: OAuth browser redirects and magic link callbacks
# - http.server: Local callback server for OAuth authorization code exchange
# - socket: Network operations for OAuth callback server
# - threading: Background processing for non-blocking OAuth flow

# System monitoring (for health checks)
psutil>=5.9.0

# Development tools (optional)
pytest>=7.0.0
pytest-qt>=4.0.0
black>=22.0.0
flake8>=4.0.0

# Schema validation
jsonschema>=4.0.0